Extract GraphQL Tracker and call it only when needed
